Changes 1

+++ src/main/java/com/auction/biddingsystem/model/Bid.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.Data;
 
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -13,7 +14,7 @@ public class Bid {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Double bidAmount;
+    private BigDecimal bidAmount;
 
     private LocalDateTime bidTime;
